requirements.txt
numpy>=1.20.0
 torch>=1.7.1
 torchvision>=0.8.2
 opencv-python>=4.5.0
 Pillow>=8.0.0
 tqdm>=4.50.0
 scipy>=1.6.0